[distrMod] _only_ in branch 2.8:
(a) "cleverer parsing of dots" in M[L,C,D]Estimator:

Triggered by an error observed by Kornelius Rohmeyer, we now parse the dots (...) argument 
of M[L,C,D]Estimator a  little more closely and filter out obvious clashes: We filter out taboo arguments 
(causing clashes with formals in optim/optimize and calling arguments); from the remaining ones only 
[exactly] _named_ arguments of the optimizer (optim/optimize) and those matching either exactly arguments 
of the criterion or all remaining ones (if ... is a formal of the criterion) are passed on: 
This way, one can, if desired, do evaluation-wise validity checks in the criterion. 

(b) for diagnostic purposes, MCEstimate-class gains a slot optimReturn (of class "ANY" and filled by
NULL by default) which is filled by the return value of the optimizer in "mceCalc" -- it has a
corresponding accessor

As this is more for internal purposes, example code is in lines ll 256--294 in mleCalc-methods.R
(wrapped in a if(FALSE){ }).  


git-svn-id: svn://svn.r-forge.r-project.org/svnroot/distr@1232 edb9625f-4e0d-4859-8d74-9fd3b1da38cb